Move `utils.address.tests` in their own file
